+++ aws/deploy-kops-cluster.sh
@@ -71,7 +71,7 @@ create_kops_iam_user() {
 	)
 
 	# Create user and group
-	if ! aws iam list-users | grep -q "UserName.*$PROFILE"; then
+	if ! aws iam get-user --user $PROFILE &> /dev/null; then
 		aws iam create-group --group-name $PROFILE || return $?
 		# Attach policies to group
 		for p in ${policy[@]}; do
@@ -85,41 +85,44 @@ create_kops_iam_user() {
 	fi
 
 	# Create Access Key ID and Secret Access Key
-	local tmp=$(mktemp)
-	if ! aws iam create-access-key --user-name $PROFILE > $tmp; then
-		local ret=$?
+	if ! grep -q $PROFILE $HOME/.aws/credentials; then
+		local tmp=$(mktemp)
+		if ! aws iam create-access-key --user-name $PROFILE > $tmp; then
+			local ret=$?
+			rm $tmp
+			return $ret
+		fi
+		local access_key=$(jq -r .AccessKey.AccessKeyId < $tmp)
+		local secret_key=$(jq -r .AccessKey.SecretAccessKey < $tmp)
+
+		# Populate AWS credentials
+		cat <<- _CREDENTIALS >> $HOME/.aws/credentials
+
+		[$PROFILE]
+		aws_access_key_id=$access_key
+		aws_secret_access_key=$secret_key
+		_CREDENTIALS
+
+		# Populate AWS config
+		cat <<- _CONFIG >> $HOME/.aws/config
+
+		[profile $PROFILE]
+		region=$REGION
+		_CONFIG
+
+		# Remove temporary AWS config file
 		rm $tmp
-		return $ret
 	fi
-	local access_key=$(jq -r .AccessKey.AccessKeyId < $tmp)
-	local secret_key=$(jq -r .AccessKey.SecretAccessKey < $tmp)
 
-	# Populate AWS credentials
-	cat <<- _CREDENTIALS >> $HOME/.aws/credentials
-
-	[$PROFILE]
-	aws_access_key_id=$access_key
-	aws_secret_access_key=$secret_key
-	_CREDENTIALS
-
-	# Populate AWS config
-	cat <<- _CONFIG >> $HOME/.aws/config
-
-	[profile $PROFILE]
-	region=$REGION
-	_CONFIG
-
-	# Remove temporary AWS config file
-	rm $tmp
-
-	# Verify kops user was successfully created
-	aws iam get-user --user $PROFILE
-	return $?
+	return 0
 }
 
 
 # An S3 bucket must be created to store the state of the k8s cluster
 create_kops_state_store() {
+	# Check if bucket is already created and if we have permission to access it
+	aws s3api head-bucket --bucket $(basename $KOPS_STATE_STORE) && return 0
+
 	local state_store=$(basename $KOPS_STATE_STORE)
 	aws s3 mb s3://$state_store || return $?
 	aws s3api put-bucket-versioning \
@@ -135,6 +138,8 @@ create_kops_state_store() {
 
 # Setup AWS Route53 to access cluster via DNS subdomain,
 create_route53_dns() {
+	aws route53 list-hosted-zones | grep -q $NAME && return 0
+
 	local id=$(uuidgen) && \
 		aws route53 create-hosted-zone --name $NAME --caller-reference $id | \
 			jq .DelegationSet.NameServers
@@ -143,7 +148,9 @@ create_route53_dns() {
 	You will need to create a new SUBDOMAIN, and use the 4 NS records received from the above command for the new SUBDOMAIN.
 	This MUST be done in order to use your cluster.
 	Do NOT change your top level NS record, or you might take your site offline.
+
 	_SUBDOMAIN
+	read -n 1 -p "Press anykey to proceed."
 	return $ret
 }
 
@@ -196,15 +203,14 @@ if ! install_dependencies; then
 fi
 
 # Verify kops user has been created
-if ! (grep -q $PROFILE $HOME/.aws/credentials || create_kops_iam_user); then
+if ! create_kops_iam_user; then
 	prompt_error "failed to generate $PROFILE user"
 	exit 1
-else
-	# awscli does not export these variables for kops to use
-	export AWS_PROFILE=$PROFILE
-	export AWS_ACCESS_KEY_ID=$(aws configure get aws_access_key_id)
-	export AWS_SECRET_ACCESS_KEY=$(aws configure get aws_secret_access_key)
 fi
+# awscli does not export these variables for kops to use
+export AWS_PROFILE=$PROFILE
+export AWS_ACCESS_KEY_ID=$(aws configure get aws_access_key_id)
+export AWS_SECRET_ACCESS_KEY=$(aws configure get aws_secret_access_key)
 
 # Check for user input
 while [ -n "$1" ]; do
@@ -249,13 +255,13 @@ while [ -n "$1" ]; do
 done
 
 # Verify kops state store has been created
-if ! (aws s3 ls | grep -q $(basename $KOPS_STATE_STORE) || create_kops_state_store); then
+if ! create_kops_state_store; then
 	prompt_error "failed to create kops state store in S3"
 	exit 1
 fi
 
 # Route53 DNS needs to be deployed to use kops unless using gossip based DNS
-#if ! (aws route53 list-hosted-zones | grep -q $NAME || create_route53_dns); then
+#if ! create_route53_dns; then
 #	prompt_error "failed to create Route53"
 #	exit 1
 #fi
